Backport 7404ddf24a162cff445cd0a26aec446461988bc8
