+++ test/hotspot/jtreg/vmTestbase/gc/g1/unloading/bytecode/BytecodeGeneratorFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,7 +26,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Random;
 
-import vm.share.InMemoryJavaCompiler;
+import jdk.test.lib.compiler.InMemoryJavaCompiler;
 
 /**
  * BytecodeFactory that employs in memory compilation.
@@ -44,12 +44,10 @@ public class BytecodeGeneratorFactory implements BytecodeFactory {
 
     @Override
     public Bytecode createBytecode(String className) {
-        Map<String, CharSequence> sources = new HashMap<String, CharSequence>();
-        sources.put(className, sourceGenerator.generateSource(className,
+        byte[] bytecode = InMemoryJavaCompiler.compile(className, sourceGenerator.generateSource(className,
                 "public static void main() { System.out.println(\"From main method in in-mem-compiled code " + random.nextGaussian() +
                         " + str_bytesToReplace0 str_bytesToReplace1\"); }\n " +
                         "public static int methodForCompilation(Object object) { int i = object.hashCode(); i = i * 2000 / 1994 + 153; return i; }\n"));
-        byte[] bytecode = InMemoryJavaCompiler.compile(sources).values().iterator().next();
         return new Bytecode(className, bytecode);
     }
