+++ test/lib/jdk/test/lib/compiler/InMemoryJavaCompiler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,11 +26,18 @@ package jdk.test.lib.compiler;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.io.StringWriter;
+import java.io.Writer;
 
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
 import javax.tools.FileObject;
 import javax.tools.ForwardingJavaFileManager;
@@ -76,36 +83,6 @@ import javax.tools.ToolProvider;
  * </pre>
  */
 public class InMemoryJavaCompiler {
-    private static class MemoryJavaFileObject extends SimpleJavaFileObject {
-        private final String className;
-        private final CharSequence sourceCode;
-        private final ByteArrayOutputStream byteCode;
-
-        public MemoryJavaFileObject(String className, CharSequence sourceCode) {
-            super(URI.create("string:///" + className.replace('.','/') + Kind.SOURCE.extension), Kind.SOURCE);
-            this.className = className;
-            this.sourceCode = sourceCode;
-            this.byteCode = new ByteArrayOutputStream();
-        }
-
-        @Override
-        public CharSequence getCharContent(boolean ignoreEncodingErrors) {
-            return sourceCode;
-        }
-
-        @Override
-        public OutputStream openOutputStream() throws IOException {
-            return byteCode;
-        }
-
-        public byte[] getByteCode() {
-            return byteCode.toByteArray();
-        }
-
-        public String getClassName() {
-            return className;
-        }
-    }
 
     private static class FileManagerWrapper extends ForwardingJavaFileManager<JavaFileManager> {
         private static final Location PATCH_LOCATION = new Location() {
@@ -119,12 +96,13 @@ public class InMemoryJavaCompiler {
                 return false;
             }
         };
-        private final MemoryJavaFileObject file;
+        private final SourceFile srcFile;
+        private ClassFile clsFile;
         private final String moduleOverride;
 
-        public FileManagerWrapper(MemoryJavaFileObject file, String moduleOverride) {
+        public FileManagerWrapper(SourceFile file, String moduleOverride) {
             super(getCompiler().getStandardFileManager(null, null, null));
-            this.file = file;
+            this.srcFile = file;
             this.moduleOverride = moduleOverride;
         }
 
@@ -132,16 +110,17 @@ public class InMemoryJavaCompiler {
         public JavaFileObject getJavaFileForOutput(Location location, String className,
                                                    Kind kind, FileObject sibling)
             throws IOException {
-            if (!file.getClassName().equals(className)) {
-                throw new IOException("Expected class with name " + file.getClassName() +
+            if (!srcFile.getClassName().equals(className)) {
+                throw new IOException("Expected class with name " + srcFile.getClassName() +
                                       ", but got " + className);
             }
-            return file;
+            clsFile = new ClassFile(className);
+            return clsFile;
         }
 
         @Override
         public Location getLocationForModule(Location location, JavaFileObject fo) throws IOException {
-            if (fo == file && moduleOverride != null) {
+            if (fo == srcFile && moduleOverride != null) {
                 return PATCH_LOCATION;
             }
             return super.getLocationForModule(location, fo);
@@ -160,6 +139,100 @@ public class InMemoryJavaCompiler {
             return super.hasLocation(location) || location == StandardLocation.PATCH_MODULE_PATH;
         }
 
+        public byte[] getByteCode() {
+            return clsFile.toByteArray();
+        }
+
+    }
+
+    // Wraper for class file
+    static class ClassFile extends SimpleJavaFileObject {
+
+        private final ByteArrayOutputStream baos = new ByteArrayOutputStream();
+
+        protected ClassFile(String name) {
+            super(URI.create("memo:///" + name.replace('.', '/') + Kind.CLASS.extension), Kind.CLASS);
+        }
+
+        @Override
+        public ByteArrayOutputStream openOutputStream() { return this.baos; }
+
+        byte[] toByteArray() { return baos.toByteArray(); }
+    }
+
+    // File manager which spawns ClassFile instances by demand
+    static class FileManager extends ForwardingJavaFileManager<JavaFileManager> {
+
+        private Map<String, ClassFile> classesMap = new HashMap<String, ClassFile>();
+
+        protected FileManager(JavaFileManager fileManager) {
+            super(fileManager);
+        }
+
+        @Override
+        public ClassFile getJavaFileForOutput(Location location, String name, JavaFileObject.Kind kind, FileObject source) {
+            ClassFile classFile = new ClassFile(name);
+            classesMap.put(name, classFile);
+            return classFile;
+        }
+
+        public Map<String, byte[]> getByteCode() {
+            Map<String, byte[]> result = new HashMap<String, byte[]>();
+            for (Entry<String, ClassFile> entry : classesMap.entrySet()) {
+                result.put(entry.getKey(), entry.getValue().toByteArray());
+            }
+            return result;
+        }
+    }
+
+    // Wrapper for source file
+    static class SourceFile extends SimpleJavaFileObject {
+
+        private CharSequence sourceCode;
+        private String className;
+
+        public SourceFile(String name, CharSequence sourceCode) {
+            super(URI.create("memo:///" + name.replace('.', '/') + Kind.SOURCE.extension), Kind.SOURCE);
+            this.sourceCode = sourceCode;
+            this.className = name;
+        }
+
+        @Override
+        public CharSequence getCharContent(boolean ignore) {
+            return this.sourceCode;
+        }
+
+        public String getClassName() {
+            return this.className;
+        }
+    }
+
+    /**
+     * Compiles the list of classes with the given map of name and source code.
+     * This overloaded version of compile is useful for batch compile use cases.
+     *
+     * @param inputMap The map containing the name of the class and corresponding source code
+     * @throws RuntimeException if the compilation did not succeed
+     * @return The resulting byte code from the compilation
+     */
+    public static Map<String, byte[]> compile(Map<String, ? extends CharSequence> inputMap) {
+        Collection<JavaFileObject> sourceFiles = new LinkedList<JavaFileObject>();
+        for (Entry<String, ? extends CharSequence> entry : inputMap.entrySet()) {
+            sourceFiles.add(new SourceFile(entry.getKey(), entry.getValue()));
+        }
+
+        JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
+        FileManager fileManager = new FileManager(compiler.getStandardFileManager(null, null, null));
+
+        Writer writer = new StringWriter();
+        Boolean exitCode = compiler.getTask(writer, fileManager, null, null, null, sourceFiles).call();
+        if (!exitCode) {
+            System.out.println("*********** javac output begin ***********");
+            System.out.println(writer.toString());
+            System.out.println("*********** javac output end ***********");
+            throw new RuntimeException("Test bug: in memory compilation failed.");
+        }
+        return fileManager.getByteCode();
     }
 
     /**
@@ -173,7 +246,7 @@ public class InMemoryJavaCompiler {
      * @return The resulting byte code from the compilation
      */
     public static byte[] compile(String className, CharSequence sourceCode, String... options) {
-        MemoryJavaFileObject file = new MemoryJavaFileObject(className, sourceCode);
+        SourceFile file = new SourceFile(className, sourceCode);
         List<String> opts = new ArrayList<>();
         String moduleOverride = null;
         for (String opt : options) {
@@ -183,13 +256,13 @@ public class InMemoryJavaCompiler {
                 opts.add(opt);
             }
         }
-        try (JavaFileManager fileManager = new FileManagerWrapper(file, moduleOverride)) {
+        try (FileManagerWrapper fileManager = new FileManagerWrapper(file, moduleOverride)) {
             CompilationTask task = getCompiler().getTask(null, fileManager, null, opts, null, Arrays.asList(file));
             if (!task.call()) {
                 throw new RuntimeException("Could not compile " + className + " with source code " + sourceCode);
             }
 
-            return file.getByteCode();
+            return fileManager.getByteCode();
         } catch (IOException ioe) {
             throw new RuntimeException(ioe);
         }
